Перейменував та видалив зайві файли. Додав приклади додання нотаток та пошуку.
