+++ main_DenM.py
@@ -1,3 +1,10 @@
-from magic_ball import magic_8_ball
+from note_book import magic_8_ball, note_add, note_find
 
+# виведення відповіді магічної кулі
 print(magic_8_ball())
+
+# додання нової нотатки
+note_add('')
+
+# пошук по всім полям нотаток якщо перед словом # то пошук тільки по тегам
+note_find('')
