+++ note_book.py
@@ -2,6 +2,54 @@ from collections import UserDict
 from datetime import datetime
 import json
 import os
+import random
+
+def magic_8_ball():
+    responses = [
+        "Безперечно!",
+        "Так точно!",
+        "Вельми ймовірно",
+        "Можливо",
+        "Не можу вам сказати зараз",
+        "Спробуйте ще раз пізніше",
+        "Краще не розраховувати на це",
+        "Відповідь невизначена",
+        "Це малоймовірно",
+        "Ні, це неможливо",
+        "Духи говорять так!",
+        "Є сумніви",
+        "Думаю що ні",
+        "Шанси хороші",
+        "Духи говорять ні",
+        "Без сумнівів",
+        "Шансів мало",
+        "Звичайно, якщо хочеш стати пінгвіном.",
+        "Тільки якщо готовий танцювати під дощем.",
+        "Звичайно, але пам'ятай, що це не змінить твій настрій.",
+        "Запитай пізніше. Зараз я зайнята готуванням чарівного пирога.",
+        "Ймовірно так! Я бачила це в своєму кришталевому шарі. :-)",
+        "Ніколи не кажи ніколи. Хіба, що магічна куля принесе тобі піцу.",
+        "Чи не краще тобі поговорити з моїм колегою, Магічнимм Барменом? Він спеціаліст з усіх важливих справ.",
+        "Спитай у кішки. Вона знає багато секретів всесвіту.",
+        "Завжди пам'ятайте: краще питати дурні питання, ніж дурно відповідати на питання. Я готова відповісти на будь-яке!",
+        "Відповідь закодована в потоці нейронних синапсів... Але нажаль, у мене не залучення до інтернету.",
+        "Впевнена на 99,9%... але є ще цей 0,1%, коли всі правила перестають діяти.",
+        "Знаєте, іноді питання без відповідей - найцікавіші. Так, ваше питання може стати наступним кандидатом на цю нагороду!",
+        "Це випадково, але відповідь - так!",
+        "На жаль, магічна куля відпочиває на Гавайських островах, але я можу залишити тобі свої 27 смішних порад!",
+        "Ох, це складне питання. Магічна куля вирішила пожартувати - питай знову!",
+        "Не можу відповісти прямо зараз, я ще переглядаю ""Курс магічних відповідей для чайників"".",
+        "Погодься на піцу і пиво, і відповідь буде - так!",
+        "Запитання надто складне для магічної кулі, спробуй з кришталевим шаром.",
+        "Магічна куля бачить велике щастя у твоєму майбутньому - але не знає, чи це через відповідь на це питання.",
+        "Питання настільки хороше, що магічна куля вирішила сміятись і не відповідати.",
+        "Ні, але якщо принесеш мені каву, то можливо згодом зміню думку.",
+        "Зірки кажуть - можливо!",
+        "Зроби паузу, випий чашку гарячого шоколаду і подивись комедійний серіал - це теж розвага!",
+        "Магічна куля каже, що можливість бути в гарному настрої - це найкращий депозит, який можеш мати!"
+    ]
+    random.shuffle(responses)
+    return random.choice(responses)
 
 class CustomException(Exception):
     pass
